extract primitives
