+++ chains/node/cli/Cargo.toml
@@ -57,6 +57,7 @@ fp-storage						= { default-features = false, git = "https://github.com/parityte
 # local packages
 birthday-runtime = { package = "birthday-runtime", path = "../../runtime/birthday" }
 meta-runtime = { package = "meta-runtime", path = "../../runtime/meta" }
+meta-primitives = { package = "meta-primitives", path = "../../primitives" }
 
 [build-dependencies]
 substrate-build-script-utils = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.25" }
