+++ chains/node/cli/src/chain_spec/meta.rs
@@ -1,9 +1,13 @@
-use meta_runtime::{AccountId, GenesisConfig, Signature, WASM_BINARY};
+use meta_runtime::{ GenesisConfig, WASM_BINARY};
+use meta_primitives::{AccountId, Signature};
 use sc_service::ChainType;
 use sp_core::{sr25519, Pair, Public, H160, U256};
 use sp_runtime::traits::{IdentifyAccount, Verify};
 use std::{collections::BTreeMap, str::FromStr};
 use crate::chain_spec::{ChainSpec, get_account_id_from_seed};
+
+/// Specialized `ChainSpec`. This is a specialization of the general Substrate ChainSpec type.
+pub type ChainSpec = sc_service::GenericChainSpec<GenesisConfig>;
 
 pub fn development_config() -> Result<ChainSpec, String> {
 	let wasm_binary = WASM_BINARY.ok_or_else(|| "Development wasm not available".to_string())?;
