+++ chains/runtime/meta/Cargo.toml
@@ -40,6 +40,8 @@ pallet-base-fee 							= { default-features = false, git = "https://github.com/p
 pallet-dynamic-fee 							= { default-features = false, git = "https://github.com/paritytech/frontier.git", branch = "polkadot-v0.9.25" }
 pallet-ethereum 							= { default-features = false, git = "https://github.com/paritytech/frontier.git", branch = "polkadot-v0.9.25" }
 pallet-evm 									= { default-features = false, git = "https://github.com/paritytech/frontier.git", branch = "polkadot-v0.9.25" }
+## Local
+meta-primitives								= { default-features = false, path = "../../primitives"}
 
 [build-dependencies]
 substrate-wasm-builder 						= { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.25" }
